Replaces rTorrent check with Flood UI check
Updates the healthcheck to verify the Flood UI is accessible
This prevents hammering the rtorrent RPC interface which can lead to the container hanging
and responsive instead of directly connecting to rTorrent.

This change ensures the healthcheck reflects the availability
of the user interface, which is a more relevant indicator of
the application's overall health.

Also makes Flood UI a critical check after startup period.